ui changes in landing
